Fixing the new restructuring causing break in the build of library and breaking DPCPP compiler.
