+++ benchmark/clBench/acl/CMakeLists.txt
@@ -6,8 +6,8 @@ endif()
 
 function(generate_acl_benchmark name source definition)
   add_executable(${name} ${source} main.cpp)
-  target_link_libraries(${name} PRIVATE benchmark acl sycl_blas_common sycl_blas Clara::Clara)
-  target_include_directories(${name} PRIVATE ${CBLAS_INCLUDE})
+  target_link_libraries(${name} PRIVATE benchmark acl sycl_blas Clara::Clara)
+  target_include_directories(${name} PRIVATE ${CBLAS_INCLUDE} ${BLAS_BENCH} ${SYCLBLAS_INCLUDE} ${SYCLBLAS_COMMON_INCLUDE_DIR})
   target_compile_definitions(${name} PRIVATE ${definition})
 
   if(BLAS_VERIFY_BENCHMARK)
