+++ benchmark/clBench/clblas/blas3/trsm.cpp
