+++ benchmark/clBench/clblast/CMakeLists.txt
@@ -25,8 +25,8 @@ set(sources
 foreach(clblast_benchmark ${sources})
   get_filename_component(clblast_bench_exec ${clblast_benchmark} NAME_WE)
   add_executable(bench_clblast_${clblast_bench_exec} ${clblast_benchmark} main.cpp)
-  target_link_libraries(bench_clblast_${clblast_bench_exec} PRIVATE benchmark Clara::Clara clblast sycl_blas_common)
-  target_include_directories(bench_clblast_${clblast_bench_exec} PRIVATE ${CBLAS_INCLUDE})
+  target_link_libraries(bench_clblast_${clblast_bench_exec} PRIVATE benchmark Clara::Clara clblast)
+  target_include_directories(bench_clblast_${clblast_bench_exec} PRIVATE ${SYCLBLAS_INCLUDE} ${CBLAS_INCLUDE} ${BLAS_BENCH} ${SYCLBLAS_COMMON_INCLUDE_DIR})
 
   if(BLAS_VERIFY_BENCHMARK)
     target_link_libraries(bench_clblast_${clblast_bench_exec} PRIVATE blas::blas)
