+++ benchmark/clBench/clblast/blas1/asum.cpp
