+++ benchmark/clBench/clblast/blas1/dot.cpp
