+++ benchmark/clBench/clblast/blas1/iamax.cpp
