+++ benchmark/clBench/clblast/blas1/iamin.cpp
