+++ benchmark/clBench/clblast/blas1/nrm2.cpp
