+++ benchmark/clBench/clblast/blas1/scal.cpp
