+++ benchmark/clBench/clblast/blas3/gemm_batched.cpp
