+++ test/unittest/CMakeLists.txt
@@ -73,8 +73,8 @@ foreach(blas_test ${SYCL_UNITTEST_SRCS})
   if(STRESS_TESTING)
     target_compile_definitions(${test_exec} PRIVATE STRESS_TESTING)
   endif()
-  target_link_libraries(${test_exec} PRIVATE gtest_main Clara::Clara blas::blas sycl_blas_common sycl_blas)
-  target_include_directories(${test_exec} PRIVATE ${CBLAS_INCLUDE})
+  target_link_libraries(${test_exec} PRIVATE gtest_main Clara::Clara blas::blas sycl_blas)
+  target_include_directories(${test_exec} PRIVATE ${CBLAS_INCLUDE} ${SYCLBLAS_COMMON_INCLUDE_DIR})
   if(TEST_DEVICE)
     add_test(NAME ${test_exec} COMMAND ${CMAKE_CURRENT_BINARY_DIR}/${test_exec} --device ${TEST_DEVICE})
   else()
